+++ memiavl/snapshot_test.go
@@ -12,7 +12,7 @@ import (
 
 func TestSnapshotEncodingRoundTrip(t *testing.T) {
 	// setup test tree
-	tree := NewEmptyTree(0)
+	tree := New(0)
 	for _, changes := range ChangeSets[:len(ChangeSets)-1] {
 		_, _, err := tree.ApplyChangeSet(changes, true)
 		require.NoError(t, err)
@@ -24,7 +24,7 @@ func TestSnapshotEncodingRoundTrip(t *testing.T) {
 	snapshot, err := OpenSnapshot(snapshotDir)
 	require.NoError(t, err)
 
-	tree2 := NewFromSnapshot(snapshot, true)
+	tree2 := NewFromSnapshot(snapshot, true, 0)
 
 	require.Equal(t, tree.Version(), tree2.Version())
 	require.Equal(t, tree.RootHash(), tree2.RootHash())
@@ -40,7 +40,7 @@ func TestSnapshotEncodingRoundTrip(t *testing.T) {
 	// test modify tree loaded from snapshot
 	snapshot, err = OpenSnapshot(snapshotDir)
 	require.NoError(t, err)
-	tree3 := NewFromSnapshot(snapshot, true)
+	tree3 := NewFromSnapshot(snapshot, true, 0)
 	hash, v, err := tree3.ApplyChangeSet(ChangeSets[len(ChangeSets)-1], true)
 	require.NoError(t, err)
 	require.Equal(t, RefHashes[len(ChangeSets)-1], hash)
@@ -60,7 +60,7 @@ func TestSnapshotExport(t *testing.T) {
 	}
 
 	// setup test tree
-	tree := NewEmptyTree(0)
+	tree := New(0)
 	for _, changes := range ChangeSets[:3] {
 		_, _, err := tree.ApplyChangeSet(changes, true)
 		require.NoError(t, err)
@@ -88,7 +88,7 @@ func TestSnapshotExport(t *testing.T) {
 
 func TestSnapshotImportExport(t *testing.T) {
 	// setup test tree
-	tree := NewEmptyTree(0)
+	tree := New(0)
 	for _, changes := range ChangeSets {
 		_, _, err := tree.ApplyChangeSet(changes, true)
 		require.NoError(t, err)