+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 
 - [#833](https://github.com/crypto-org-chain/cronos/pull/833) Fix rollback command.
 - [#945](https://github.com/crypto-org-chain/cronos/pull/945) Fix no handler exists for proposal type error when update-client due to wrong ibc route.
+- [#1036](https://github.com/crypto-org-chain/cronos/pull/1036) Fix memiavl import memory leak.
 
 ### Improvements