+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - [#986](https://github.com/crypto-org-chain/cronos/pull/986) Use go 1.20.
 - [#984](https://github.com/crypto-org-chain/cronos/pull/984) experimental integration of memiavl.
 - [#985](https://github.com/crypto-org-chain/cronos/pull/985) Fix versiondb verify command on older versions
+- [#1038](https://github.com/crypto-org-chain/cronos/pull/1038) Update ibc-go to `v5.2.1`.
 
 *April 13, 2023*